Integrate RSpec for testing in application
RSpec was added to the application for tests. Configuration files - .rspec, rails_helper.rb, and spec_helper.rb, have been set up for its use. The rspec-rails gem was included in the Gemfile and Gemfile.lock. This integration allows the automatic creation of spec files corresponding to new classes and supports a wide range of dev testing approaches.