Implement GraphQL across application
The application has been updated to use GraphQL, with the creation of base classes for various GraphQL types, mutations, and query fields. A GraphQL controller was added to handle incoming GraphQL requests. The integration of Relay specification is handled through the inclusion of Connection and Edge behaviours in appropriate classes. GraphQL is also added to the Gemfile for dependency tracking. This change moves the application to adopt GraphQL for a flexible, efficient data query and manipulation.